Support more recent ActiveFedora usage
This PR was driven by the ETD work cycle in early 2018, the goal of which was to
update the `hydra_etd` app such that it uses somewhat modern dependencies. Part of
this work was upgrading the app's ActiveFedora usage from 2.x to the latest 8.x
version. In this AF version, relations like `belongs_to` and `has_many` expect their
first argument to be a symbol rather than a string.

Additionally:

* `class_name` values including the `EtdModel` namespace caused inverse reflections to break in AF. You can see this was patched around in etd-robots earlier, and now neither etd-robots or hydra_etd will need local patches to deal with this. See https://github.com/sul-dlss/etd-robots/commit/8052ffea2c7496510d1fe73a842e276abff32fb8#diff-b5d01edc1f6ec3519ee5dc5c4dbc235c
* `workflow` and `properties` datastreams should be unversioned by default
* `DC` datastream should be the Fedora default, which is based on DC11 rather than DCTERMS. This datastream is already part of the AF/Rubydora stack and does not need to be explicitly declared